% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_raters}
\alias{compare_raters}
\title{Compare Two Raters' Codes Cell-by-Cell}
\usage{
compare_raters(
  rater_1,
  rater_2,
  beg_symbol = "",
  sep_symbol = "/",
  output_class = NULL
)
}
\arguments{
\item{rater_1}{Rater 1's data. Can be a \code{matrix}, \code{data.frame}, or \code{tibble}.}

\item{rater_2}{Rater 2's data. Can be a \code{matrix}, \code{data.frame}, or \code{tibble}.}

\item{beg_symbol}{(Optional) text string appearing before coder 1's code for inconsistent cells.}

\item{sep_symbol}{Text string separating both coders' codes for inconsistent cells.}

\item{output_class}{(Optional) text string specifying what to output data as. Can be one of: "matrix", "data.frame", or "tibble". By default, the outputted data will be the same format as \code{rater_1}.}
}
\value{
A data frame (if input is a \code{data.frame}), a tibble (if input is a \code{tibble}), or a matrix (if input is a \code{matrix}).
}
\description{
Compares two matrices, data frames, or tibbles of rater codes, returning the same structure with cell-level differences marked using \code{beg_symbol} and \code{sep_symbol}.
}
\examples{
X_dat <- data.frame(var_1 = c(1, 3, 2),
                   var_2 = c(2, 2, 2),
                   var_3 = c(NA, 1, 1))
Y_dat <- data.frame(var_1 = c(1, 3, 2),
                   var_2 = c(2, 2, 1),
                   var_3 = c(3, 1, 0))
compare_raters(X_dat, Y_dat)

# Can use different symbols than \ to separate two raters' codes.
# Separate different scores with "or":
compare_raters(X_dat, Y_dat,
               sep_symbol = " or ")

# Separate different scores with "Rater 1:" and "Rater 2:"
compare_raters(X_dat, Y_dat,
               beg_symbol = "Rater 1: ",
               sep_symbol = "; Rater 2: ")
}
