% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_vectors}
\alias{compare_vectors}
\title{Compare All Elements from Two Vectors and Flag Inconsistencies}
\usage{
compare_vectors(vec1, vec2, beg_symbol = "", sep_symbol = "/")
}
\arguments{
\item{vec1}{First vector (same length as vec2).}

\item{vec2}{Second vector (same length as vec1).}

\item{beg_symbol}{(Optional) text string appearing before vec1's entry for inconsistent elements.}

\item{sep_symbol}{Text string separating both vectors' values for inconsistent elements.}
}
\value{
A character vector summarizing any inconsistencies between vectors.
}
\description{
Compares elements of two vectors element-wise, flagging any inconsistency with
a symbol denoted with sep_symbol.
}
\note{
This function compares two vectors and returns a vector of compared entries.
If you have a more full data set with two coders' entries across several variables,
please use gcgenie::compare_raters().
}
\examples{
# With character inputs
char_1 <- c("yes", "yes", "no", "yes", "no")
char_2 <- c("no", "yes", "no", "no", "yes")
compare_vectors(char_1, char_2)

# With numeric inputs
num_1 <- 1:5
num_2 <- 5:9
compare_vectors(num_1, num_2, sep_symbol = ";")

# In tidy context
# Suppose
df <- data.frame(
    rater_1 = char_1,
    rater_2 = char_2
)

df |>
    dplyr::mutate(compare_col = compare_vectors(rater_1, rater_2))
# With numeric inputs
compare_vectors(1:3, 3:5)
}
