% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_functions.R
\name{count_plot}
\alias{count_plot}
\title{Generate a Count Plot from a Count Table}
\usage{
count_plot(
  dat,
  var = NULL,
  count_var = n,
  head_n = nrow(dat),
  label = FALSE,
  hjust = -0.1,
  color = "black",
  fill = "royalblue"
)
}
\arguments{
\item{dat}{Frequency table (tibble, data.frame, or matrix) that includes a categorical (counted or grouped) and a corresponding count column (usually, 'n').}

\item{var}{The categorical variable in the frequency table, unquoted.}

\item{count_var}{The count of the variable, usually 'n', unqouted. By default, it is assumed to be 'n'.}

\item{head_n}{The number of final unique values to include in the plot.}

\item{label}{(Optional) If TRUE, will provide text of the frequnecy count to the right of each bar.}

\item{hjust}{(Optional) Horizontal offset for x-axis labels; effective only when label = TRUE.}

\item{color}{(Optional) Color of the bar outlines.}

\item{fill}{(Optional) Color for the inside the bars.}
}
\value{
A ggplot2 column plot of frequencies.
}
\description{
Generates a column plot from a frequency table,
with categories sorted by descending frequency.
}
\examples{

df <- data.frame(person = c("Jane,Joe", "Joe", "Joe,Kai", "Kai"))

# Regular count plot without labels
df |> unique_count(person) |> count_plot(person)
# With labels
df |> unique_count(person) |> count_plot(person, label = TRUE, hjust = 2)
}
