% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_functions.R
\name{freq_count}
\alias{freq_count}
\title{Convenience wrapper to \code{dplyr::count()}}
\usage{
freq_count(dat, ..., sort = TRUE)
}
\arguments{
\item{dat}{Data set as a matrix, data frame, or tibble.}

\item{...}{Additional arguments to add to dplyr::count, such as the variables to count.}

\item{sort}{(optional) when set to TRUE, sorts the output
so that higher-frequency items appear above less frequent ones.
Defaults to TRUE.}
}
\value{
Returns an object of the same class as dat: a tibble, data frame, or matrix.
}
\description{
Like dplyr::count(), but always sorts descending and adds a perc_n column with percentage of total counts.
}
\examples{
dat <- data.frame(
    name = c("John", "Jane", "Jim", "Jill", "Joe", "Anna"),
    occupation = c("Academic", "Academic", "Librarian", "Firefighter", "Doctor", "Doctor"),
    native_language = c("En", "En", "Fr", "Fr", "En", "Fr"))

dat |> freq_count(occupation)
dat |> freq_count(native_language)
dat |> freq_count(occupation, native_language)
}
