% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_customs.R
\name{geom_histogram_custom}
\alias{geom_histogram_custom}
\title{Custom ggplot2 histogram}
\usage{
geom_histogram_custom(..., color = "black", fill = "royalblue")
}
\arguments{
\item{...}{Arguments passed to ggplot2::geom_histogram().}

\item{color}{(Optional) Color of the outline of each bar.}

\item{fill}{(Optional) Color of the bars themselves.}
}
\value{
A histogram plot.
}
\description{
ggplot2::goem_histogram() with useful defaults.
}
\examples{
sample_dat <- data.frame(x = rnorm(10000, mean = 10, 1))
# Regular plot
sample_dat |>
    ggplot2::ggplot(ggplot2::aes(x)) +
    geom_histogram_custom(binwidth = 0.5)

# Alter color and fill
sample_dat |>
    ggplot2::ggplot(ggplot2::aes(x)) +
    geom_histogram_custom(binwidth = 0.5, color = "grey", fill = "gold")

}
