% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_functions.R
\name{score}
\alias{score}
\title{Efficiently Score Test Data}
\usage{
score(responses, answers, display = "sum", show_questions = FALSE)
}
\arguments{
\item{responses}{Data set (tibble, data.frame, or matrix) with responses: columns are items, rows are respondents.}

\item{answers}{Vector of correct answers, with the ith element matching the ith column of the response data.}

\item{display}{(Optional) Character string specifying score display per participant: "sum" (total correct), "prop" (proportion correct), or "perc" (percentage correct).}

\item{show_questions}{(Optional) Logical. If TRUE, displays scores per question instead of per participant. Helps check if questions are being scored correctly.}
}
\value{
A numeric vector.
}
\description{
Scoring utility function that takes a data frame of responses and a
vector of correct answers, returning overall test scores per
participant or per question.
}
\note{
If a participant does not answer a question (i.e., has \code{NA} in a given cell),
the function assumes the question was answered incorrectly. This function
is stricter than \link{score_embed}
}
\examples{
responses <- data.frame(
q1 = c(NA, "b", "b", "b"),
q2 = c("d", NA, "d", "a"),
q3 = c("c", "d", "c", "a")
)

answers <- c("b", "d", "c")

# Display scores per participant
responses |> score(answers)
responses |> score(answers, display = "prop")
responses |> score(answers, display = "perc")

# Display scores per question
responses |> score(answers, show_questions = TRUE)
responses |> score(answers, display = "prop", show_questions = TRUE)
responses |> score(answers, display = "perc", show_questions = TRUE)
}
