% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_functions.R
\name{score_embed}
\alias{score_embed}
\title{Embed Scored Data into a Data Set}
\usage{
score_embed(
  responses,
  answers,
  cols = dplyr::everything(),
  display = "sum",
  name = "score"
)
}
\arguments{
\item{responses}{Data set (tibble, data.frame, or matrix) with responses: columns are items, rows are respondents.}

\item{answers}{Vector of correct answers, with the ith element matching the ith column of the response data.}

\item{cols}{A specification (e.g., q1:q10, starts_with("q")) indicating which columns contain responses to be scored. By default, all columns are assumed to contain responses to be scored.}

\item{display}{(Optional) Character string specifying score display per participant: "sum" (total correct), "prop" (proportion correct), or "perc" (percentage correct).}

\item{name}{(Optional) Character string specifying what to call the new score column.}
}
\value{
Returns an object of the same class as 'responses': a tibble, data frame, or matrix.
}
\description{
Scoring utility function that takes a data frame of responses and a
vector of correct answers, embedding the overall test scores as a new
column within the inputted data set.
}
\note{
If a participant does not answer a question (i.e., has \code{NA} in a given cell),
the function assumes the question was answered incorrectly.
}
\examples{
responses <- data.frame(
    q1 = c(NA, "b", "b", "b"),
    q2 = c("d", NA, "d", "a"),
    q3 = c("c", "d", "c", "a")
)

responses_full <- data.frame(
    var1 = c(TRUE, FALSE, TRUE, TRUE),
    var2 = c(1, 4, 2.2, 6),
    q1 = c(NA, "b", "b", "b"),
    q2 = c("d", NA, "d", "a"),
    q3 = c("c", "d", "c", "a")
)

answers <- c("b", "d", "c")

# Regular score embedding
responses |> score_embed(answers)
responses |> score_embed(answers, display = "prop")
responses |> score_embed(answers, display = "perc")

# Score embedding specifying columns to be scored
responses_full |> score_embed(answers, cols = q1:q3)
responses_full |> score_embed(answers, cols = dplyr::matches("^q"))
responses_full |> score_embed(answers, "q1":"q3")
}
