% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_imbed}
\alias{score_imbed}
\title{Imbed Scored Data into a Data Set}
\usage{
score_imbed(responses, answers, display = "sum")
}
\arguments{
\item{responses}{Data set (tibble, data.frame, or matrix) with responses: columns are items, rows are respondents.}

\item{answers}{Vector of correct answers, with the ith element matching the ith column of the response data.}

\item{display}{(Optional) Character string specifying score display per participant: "sum" (total correct), "prop" (proportion correct), or "perc" (percentage correct).}
}
\value{
A tibble if 'responses' is a tibble, a data.frame if 'responses' is a data.frame, or a matrix if 'responses' is a matrix.
}
\description{
Imbed Scored Data into a Data Set
}
\note{
If a participant does not answer a question (i.e., has \code{NA} in a given cell),
the function assumes the question was answered incorrectly.
}
\examples{
responses <- data.frame(
q1 = c(NA, "b", "b", "b"),
q2 = c("d", NA, "d", "a"),
q3 = c("c", "d", "c", "a")
)

answers <- c("b", "d", "c")

responses |> score_imbed(answers)
responses |> score_imbed(answers, display = "prop")
responses |> score_imbed(answers, display = "perc")
}
