% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_customs.R
\name{theme_custom}
\alias{theme_custom}
\title{Custom Generic ggplot2 Theme}
\usage{
theme_custom()
}
\value{
A customized ggplot2 graph.
}
\description{
Customized, clean theme for any ggplot2 graph.
}
\examples{
sample_dat <- data.frame(x = rnorm(10000, mean = 10, 1))
# Altered fill, with custom theme and labels
sample_dat |>
    ggplot2::ggplot(ggplot2::aes(x)) +
    ggplot2::geom_histogram(binwidth = 0.5, fill = "royalblue", color = "black") +
    theme_custom()
}
