% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_raters.R
\name{compare_raters}
\alias{compare_raters}
\title{Compare Two Raters' Codes Cell-by-Cell}
\usage{
compare_raters(rater_1, rater_2, beg_symbol = "", sep_symbol = "/")
}
\arguments{
\item{rater_1}{Rater 1's data. Can be a \code{matrix}, \code{data.frame}, or \code{tibble}.}

\item{rater_2}{Rater 2's data. Can be a \code{matrix}, \code{data.frame}, or \code{tibble}.}

\item{beg_symbol}{(Optional) text string appearing before coder 1's code for inconsistent cells.}

\item{sep_symbol}{Text string separating both coders' codes for inconsistent cells.}
}
\value{
A data frame (if input is a \code{data.frame}), a tibble (if input is a \code{tibble}), or a matrix (if input is a \code{matrix}).
}
\description{
This function compares two sets of rater codes, provided as two matrices,
data frames, or tibbles. It returns a new data object in the same structure,
marking any cell-level differences between the two inputs using custom
symbols specified via \code{beg_symbol} and/or \code{sep_symbol}.
}
\examples{
X_dat <- data.frame(var_1 = c(1, 3, 2),
                   var_2 = c(2, 2, 2),
                   var_3 = c(NA, 1, 1))
Y_dat <- data.frame(var_1 = c(1, 3, 2),
                   var_2 = c(2, 2, 1),
                   var_3 = c(3, 1, 0))
compare_raters(X_dat, Y_dat)

# Can use different symbols than \ to separate two raters' codes.
# Separate different scores with "or":
compare_raters(X_dat, Y_dat,
               sep_symbol = " or ")

# Separate different scores with "Rater 1:" and "Rater 2:"
compare_raters(X_dat, Y_dat,
               beg_symbol = "Rater 1: ",
               sep_symbol = "; Rater 2: ")
}
