% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_functions.R
\name{count_plot}
\alias{count_plot}
\title{Generate a Count Plot from a Count Table}
\usage{
count_plot(dat, var = NULL, count_var = n, head_n = nrow(dat), label = FALSE)
}
\arguments{
\item{dat}{Frequency table (tibble, data.frame, or matrix) that includes a categorical (counted or grouped) and a corresponding count column (usually, 'n').}

\item{var}{The categorical variable in the frequency table, unquoted.}

\item{count_var}{The count of the variable, usually 'n', unqouted. By default, it is assumed to be 'n'.}

\item{head_n}{The number of final unique values to include in the plot.}

\item{label}{(Optional) Should the plot}
}
\value{
A ggplot2 column plot of frequencies.
}
\description{
Generate a Count Plot from a Count Table
}
\examples{
# TBD
}
