% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_count.R
\name{freq_count}
\alias{freq_count}
\title{Convenience wrapper to \code{dplyr::count()}}
\usage{
freq_count(dat, ..., sort = TRUE)
}
\arguments{
\item{dat}{Data set}

\item{...}{Additional arguments to add to count, including}

\item{sort}{An optional argument that, when set to TRUE, sorts the output
so that higher-frequency items appear above less frequent ones.
Defaults to TRUE.}
}
\value{
Returns an object of the same class as dat: a tibble, data frame, or matrix.
}
\description{
This function replicates the behavior of dplyr::count(), but with more
informative defaults: it always sorts rows in descending order of frequency
and includes a perc_n column indicating the percentage each count represents
relative to the total.
}
\examples{
dat <- data.frame(
    name = c("John", "Jane", "Jim", "Jill", "Joe", "Anna"),
    occupation = c("Academic", "Academic", "Librarian", "Firefighter", "Doctor", "Doctor"),
    native_language = c("En", "En", "Fr", "Fr", "En", "Fr")
)

dat |> freq_count(occupation)
dat |> freq_count(native_language)
dat |> freq_count(occupation, native_language)
group |> freq_count()
}
