% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_functions.R
\name{score_imbed}
\alias{score_imbed}
\title{Imbed Scored Data into a Data Set}
\usage{
score_imbed(responses, answers, cols = dplyr::everything(), display = "sum")
}
\arguments{
\item{responses}{Data set (tibble, data.frame, or matrix) with responses: columns are items, rows are respondents.}

\item{answers}{Vector of correct answers, with the ith element matching the ith column of the response data.}

\item{cols}{A specification (e.g., q1:q10, starts_with("q")) indicating which columns contain responses to be scored. By default, all columns are assumed to contain responses to be scored.}

\item{display}{(Optional) Character string specifying score display per participant: "sum" (total correct), "prop" (proportion correct), or "perc" (percentage correct).}
}
\value{
Returns an object of the same class as responses: a tibble, data frame, or matrix.
}
\description{
Scoring utility function that takes a data frame of responses and a
vector of correct answers, imbedding the overall test scores as a new
column within the inputted data set.
}
\note{
If a participant does not answer a question (i.e., has \code{NA} in a given cell),
the function assumes the question was answered incorrectly.
}
\examples{
responses <- data.frame(
    q1 = c(NA, "b", "b", "b"),
    q2 = c("d", NA, "d", "a"),
    q3 = c("c", "d", "c", "a")
)

responses_full <- data.frame(
    var1 = c(TRUE, FALSE, TRUE, TRUE),
    var2 = c(1, 4, 2.2, 6),
    q1 = c(NA, "b", "b", "b"),
    q2 = c("d", NA, "d", "a"),
    q3 = c("c", "d", "c", "a")
)

answers <- c("b", "d", "c")

# Regular score imbedding
responses |> score_imbed(answers)
responses |> score_imbed(answers, display = "prop")
responses |> score_imbed(answers, display = "perc")

# Score imbedding specifying columns to be scored
responses_full |> score_imbed(answers, cols = q1:q3)
responses_full |> score_imbed(answers, cols = dplyr::matches("^q"))
responses_full |> score_imbed(answers, "q1":"q3")
}
